requirements.txt
Flask
python-dotenv
requestsFlask-SQLAlchemy
psycopg2-binary
